// src/extensions/tsserver/protocol.d.ts
  interface FileRequestArgs {
    /**
     * The file for the request (absolute pathname required).
     */
    file: string;
    projectFileName?: string;
  }
// ...
  interface TextSpan {
    /**
     * First character of the definition.
     */
    start: Location;
    /**
     * One character past last character of the definition.
     */
    end: Location;
  }
  /**
   * Object found in response messages defining a span of text in a specific source file.
   */
  interface FileSpan extends TextSpan {
    /**
     * File containing text span.
     */
    file: string;
  }
// ...
  interface NavtoRequestArgs extends FileRequestArgs {
    /**
     * Search term to navigate to from current location; term can
     * be '.*' or an identifier prefix.
     */
    searchValue: string;
    /**
     *  Optional limit on the number of items to return.
     */
    maxResultCount?: number;
    /**
     * Optional flag to indicate we want results for just the current file
     * or the entire project.
     */
    currentFileOnly?: boolean;
    projectFileName?: string;
  }
// ...
  interface NavtoItem extends FileSpan {
    /**
     * The symbol's name.
     */
    name: string;
    /**
     * The symbol's kind (such as 'className' or 'parameterName').
     */
    kind: ScriptElementKind;
    /**
     * exact, substring, or prefix.
     */
    matchKind: string;
    /**
     * If this was a case sensitive or insensitive match.
     */
    isCaseSensitive: boolean;
    /**
     * Optional modifiers for the kind (such as 'public').
     */
    kindModifiers?: string;
    /**
     * Name of symbol's container symbol (if any); for example,
     * the class name if symbol is a class member.
     */
    containerName?: string;
    /**
     * Kind of symbol's container symbol (if any).
     */
    containerKind?: ScriptElementKind;
  }